Update build_babel
Dear authors,

"ln -s ${CWD}/lib lib" creates a symbolic link in the openbabel dir pointing to the openbabel-build/lib. If the symbolic link in openbabel already exists (e.g. from a previous partial build), the current build process fails. Is it possible that we check the existence of the symbolic link instead of a failure? Thanks!

Best,
Teng